feat: 重构相关服务和路由，更新配置,实现websocket连接

+++ biz/utils/auth.go
@@ -5,10 +5,10 @@ import (
 )
 
 // 雪花算法生成uint64的随机ID
-func GenNumId() (res uint64, err error) {
+func GenNumId() (res int64, err error) {
 	node, err := snowflake.NewNode(1)
 	if err != nil {
 		return 0, err
 	}
-	return uint64(node.Generate().Int64()), nil
+	return node.Generate().Int64(), nil
 }
